refactor runner into separate files
